OpenProcessHandle takes int32

+++ pkg/security/resolvers/process/resolver_windows.go
@@ -100,7 +100,7 @@ func (p *Resolver) AddNewEntry(pid uint32, ppid uint32, file string, commandLine
 	e := p.processCacheEntryPool.Get()
 	e.PIDContext.Pid = pid
 	e.PPid = ppid
-	processHandle := procutil.OpenProcessHandle(pid)
+	processHandle := procutil.OpenProcessHandle(int32(pid))
 	username, err := procutil.GetUsernameForProcess(processHandle)
 	if err != nil {
 		log.Debugf("Couldn't get process username %v %v", pid, err)
